chore: move library out of internal
