+++ commands/status.go
